openi/requirements.txt
fs
sorl-thumbnail
django-athumb
pil
django-mailer
django-email-confirmation
django-notification
django-social-auth
south
django-imagekit
PasteDeploy==1.3.0
Paste==1.7.5.1
AddOns==0.7
BytecodeAssembler==0.6
Cheetah==2.4.4
CherryPy==3.2.2
DecoratorTools==1.8
Django==1.4
Extremes==1.1.1
FormEncode==1.2.4
Genshi==0.6
Markdown==2.1.1
SymbolType==1.0
TGScheduler==1.6.3
TurboGears==1.5.1
TurboJson==1.3.2
configobj==4.7.2
distribute==0.6.24
dj-database-url==0.2.0
django-tastypie==0.9.11
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2
django-annoying==0.7.6
django-storages==1.1.3
WebOb==1.1.1
boto==2.5.2
certifi==0.0.8
chardet==1.0.1
neo4jrestclient